+++ api/src/main/java/org/allaymc/api/entity/component/player/EntityPlayerBaseComponent.java
@@ -7,12 +7,14 @@ import org.allaymc.api.client.storage.PlayerData;
 import org.allaymc.api.entity.component.EntityBaseComponent;
 import org.allaymc.api.form.type.CustomForm;
 import org.allaymc.api.form.type.Form;
+import org.allaymc.api.item.type.ItemType;
 import org.allaymc.api.math.location.Location3ic;
 import org.allaymc.api.scoreboard.ScoreboardViewer;
 import org.allaymc.api.world.chunk.ChunkLoader;
 import org.cloudburstmc.protocol.bedrock.data.GameType;
 import org.cloudburstmc.protocol.bedrock.data.entity.EntityFlag;
 import org.cloudburstmc.protocol.bedrock.data.skin.SerializedSkin;
+import org.jetbrains.annotations.Range;
 import org.jetbrains.annotations.UnmodifiableView;
 import org.joml.Vector3d;
 import org.joml.Vector3dc;
@@ -159,7 +161,6 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent, ChunkLoa
      * Get how long the player has been using the item, in game ticks.
      *
      * @param currentTime The current time.
-     *
      * @return How long the player has been using the item, in game ticks.
      */
     long getItemUsingInAirTime(long currentTime);
@@ -397,7 +398,6 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent, ChunkLoa
      * Get a form by its ID.
      *
      * @param id The ID of the form.
-     *
      * @return The form.
      */
     Form getForm(int id);
@@ -406,7 +406,6 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent, ChunkLoa
      * Remove a form by its ID.
      *
      * @param id The ID of the form.
-     *
      * @return The removed form.
      */
     Form removeForm(int id);
@@ -446,7 +445,6 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent, ChunkLoa
      * Check if the player can reach a block at the specified position.
      *
      * @param pos The position of the block.
-     *
      * @return Whether the player can reach the block.
      */
     default boolean canReachBlock(Vector3ic pos) {
@@ -457,7 +455,6 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent, ChunkLoa
      * Check if the player can reach a point at the specified coordinates.
      *
      * @param pos The position of the point.
-     *
      * @return Whether the player can reach the point.
      */
     default boolean canReach(Vector3dc pos) {
@@ -470,7 +467,6 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent, ChunkLoa
      * @param x The x coordinate of the point.
      * @param y The y coordinate of the point.
      * @param z The z coordinate of the point.
-     *
      * @return Whether the player can reach the point.
      */
     default boolean canReach(double x, double y, double z) {
@@ -552,5 +548,51 @@ public interface EntityPlayerBaseComponent extends EntityBaseComponent, ChunkLoa
     @Override
     default boolean willBeSaved() {
         return false;
+    }
+
+    /**
+     * Sets cool down for a specific category.
+     *
+     * @param category the category to set.
+     * @param duration the cool down tick.
+     * @param send     whether send packet to the client.
+     */
+    void setCooldown(String category, @Range(from = 0, to = Integer.MAX_VALUE) int duration, boolean send);
+
+    /**
+     * @see #setCooldown(String, int, boolean)
+     */
+    default void setCooldown(String category, @Range(from = 0, to = Integer.MAX_VALUE) int duration) {
+        // NOTICE: No need to send PlayerStartItemCooldownPacket to the client since the
+        // client will display cooldown automatically if the item/category has cool down
+        setCooldown(category, duration, false);
+    }
+
+    /**
+     * Sets cool down for a specific item type.
+     *
+     * @param itemType the item type to set.
+     * @param duration the cool down tick.
+     */
+    default void setCooldown(ItemType<?> itemType, @Range(from = 0, to = Integer.MAX_VALUE) int duration) {
+        setCooldown(itemType.getIdentifier().toString(), duration);
+    }
+
+    /**
+     * Checks if the cooldown for specific category has ended.
+     *
+     * @param category the category to check.
+     * @return {@code true} if the cooldown has ended, {@code false} otherwise.
+     */
+    boolean isCooldownEnd(String category);
+
+    /**
+     * Checks if the cooldown for specific item type has ended.
+     *
+     * @param itemType the item type to check.
+     * @return {@code true} if the cooldown has ended, {@code false} otherwise.
+     */
+    default boolean isCooldownEnd(ItemType<?> itemType) {
+        return isCooldownEnd(itemType.getIdentifier().toString());
     }
 }