+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemEnderPearlStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemProjectileComponent;
 
-public interface ItemEnderPearlStack extends ItemStack {
+public interface ItemEnderPearlStack extends ItemStack, ItemProjectileComponent {
 }